requirement.txt
streamlit
pandas
plotly
pillow
numpy
scipy
